+++ NEWS.md
@@ -1,5 +1,8 @@
 # ggplot2 (development version)
 
+* `geom_ribbon()` can have varying `fill` or `alpha` in linear coordinate
+  systems (@teunbrand, #4690)
+
 # ggplot2 3.5.0
 
 This is a minor release that turned out quite beefy. It is focused on